% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns-refresh.R
\name{refresh_publc_nameservers_list}
\alias{refresh_publc_nameservers_list}
\title{Refresh the list of valid public nameservers}
\usage{
refresh_publc_nameservers_list()
}
\description{
Refresh the list of valid public nameservers
}
