% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdns-exec.R
\name{zdns_exec}
\alias{zdns_exec}
\alias{zdns_help}
\title{Raw interface to zdns}
\usage{
zdns_exec(args = c(), stdout = "", stdin = "")

zdns_help(x)
}
\description{
Pass in command-line arguments via \code{args}. Not a recommended function
unless you \emph{really} know what you're doing. Run \code{zdns_help()}
to see a list of options.
}
