% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdnsr-package.R
\docType{package}
\name{zdnsr}
\alias{zdnsr}
\alias{zdnsr-package}
\title{Perform Bulk 'DNS' Queries Using 'zdns'}
\description{
Provides wrapper/helper methods for executing 'zdns'
(\url{https://github.com/zmap/zdns}) bulk queries along with utility methods to
retrieve and cache 'Public DNS' (\url{http://public-dns.info/}) nameserver lists.
}
\details{
\itemize{
\item \code{zdns}: \url{https://github.com/zmap/zdns}
\item \code{Public DNS Lists}: \url{http://public-dns.info/}
\item URL: \url{https://gitlab.com/hrbrmstr/zdnsr}
\item BugReports: \url{https://gitlab.com/hrbrmstr/zdnsr/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
