% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdnsr-package.R
\docType{package}
\name{zdnsr}
\alias{zdnsr}
\alias{zdnsr-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/zdnsr}
\item BugReports: \url{https://gitlab.com/hrbrmstr/zdnsr/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
